внесла правки в соответствии с комментариями: добавила кастомные исключения, дто, логика инпутов в отдельный класс UserMenuHandler

+++ src/main/java/org/example/service/impl/UserServiceImpl.java
@@ -2,6 +2,9 @@ package org.example.service.impl;
 
 import org.example.dao.UserDao;
 import org.example.dao.impl.UserDaoImpl;
+import org.example.dto.CreateUserDto;
+import org.example.dto.UpdateUserDto;
+import org.example.exception.UserNotFoundException;
 import org.example.model.User;
 import org.example.service.UserService;
 
@@ -13,7 +16,9 @@ public class UserServiceImpl implements UserService {
     private final UserDao userDao = new UserDaoImpl();
 
     @Override
-    public User create(User user) {
+    public User create(CreateUserDto dto) {
+        dto.validate();
+        User user = new User(null, dto.getName(), dto.getEmail(), dto.getAge(), null);
         return userDao.create(user);
     }
 
@@ -28,12 +33,19 @@ public class UserServiceImpl implements UserService {
     }
 
     @Override
-    public User update(User user) {
+    public User update(Long id, UpdateUserDto dto) {
+        dto.validate();
+        User user = userDao.findById(id).orElseThrow(() -> new UserNotFoundException(id));
+        user.setEmail(dto.getEmail());
+        user.setAge(dto.getAge());
         return userDao.update(user);
     }
 
     @Override
     public void delete(Long id) {
+        if (userDao.findById(id).isEmpty()) {
+            throw new UserNotFoundException(id);
+        }
         userDao.delete(id);
     }
 }
